AI Improvement
In evaluation, use a function returning the steps to win to make search for
score more fine grained. Refactor and de-duplicate logic in game and put common
code in utils.py. Unit-test are extended by making two AI strategy playing
against a greedy and random strategy.

On the UI side, highlight a cell which was picked last.

Overall, the AI is doing much better. Sometimes it is a challenging opponent.
However, from time to time it makes non optimal steps.